fix(posts): score the learned rung in double precision, and label the figure honestly
Review flagged the embedding histogram's comb of spikes as float32 quantisation
in --dump. The symptom was right and the cause was a layer up: the scores are
produced by a float32 matmul over float32 embeddings, so widening only the dump
changed nothing -- 199 genuine scores still landed on 69 distinct values. The
accumulation itself is now float64, which resolves the 1e-5 spread properly and
smooths the histogram.

That is not cosmetic. bench.score breaks ties against the matcher on purpose, so
the rounding was manufacturing ties and charging them to the network: the learned
rung scored rank-1 0.030 in float32 and 0.035 in float64, with rank-5 0.111 ->
0.126 and EER 0.407 -> 0.391. Both ladder runs re-measured, the post's tables
updated, and a callout now states the precision effect where the table is
introduced rather than leaving it in a code comment.

Also from the same pass: enhance-steps.png titled its first panel "raw 500 dpi
scan" while showing the standardised 6-pixel crop, which contradicted the
minutiae argument three sections later; the lift column is a mean of per-seed
ratios and now says so, since dividing the printed means gives 2.75 where the
bottom row reads 3.76; "identical to three decimal places" described two values
differing at the third; "twice as often" was 3.6x; --controls no longer demands
the gitignored cache it never reads; and embed.py's five variants no longer
count as four.

+++ posts/fingerprint-algorithms/src/ladder.py
@@ -165,7 +165,13 @@ def learned(train, gallery, probe, epochs: int, seed: int):
     with clock("encoding and searching") as spent:
         enrolled = embed.encode(model, gallery.images)
         searched = embed.encode(model, probe.images)
-        matrix = searched @ enrolled.T
+        # The dot product in float64, not float32. Every score here lands at
+        # about 0.99999 because the embedding collapses, and float32 spacing up
+        # there is ~6e-8 -- coarse enough that 199 genuine scores round onto 69
+        # distinct values, which a 60-bin histogram of a 1e-5 range renders as a
+        # comb of spikes and gaps. The embeddings themselves stay float32, as
+        # the model produces them; only the accumulation is widened.
+        matrix = searched.astype(np.float64) @ enrolled.astype(np.float64).T
     # The loss history is reported because its *level* is the diagnostic: a
     # batch-hard triplet loss that settles at the margin has collapsed the
     # embedding. See the note at the top of embed.py.
@@ -382,9 +388,14 @@ def main() -> int:
             impostor = r.impostor
             if len(impostor) > IMPOSTOR_SAMPLE:
                 impostor = rng.choice(impostor, IMPOSTOR_SAMPLE, replace=False)
+            # float64 so the dump does not re-round scores that `learned`
+            # deliberately accumulated in double precision. Casting here was
+            # not enough on its own -- the float32 matmul had already rounded
+            # them -- but keeping it means the stored values are the ones that
+            # were scored.
             payload[f"ranks_{i}"] = r.ranks.astype(np.int16)
-            payload[f"genuine_{i}"] = r.genuine.astype(np.float32)
-            payload[f"impostor_{i}"] = impostor.astype(np.float32)
+            payload[f"genuine_{i}"] = r.genuine.astype(np.float64)
+            payload[f"impostor_{i}"] = impostor.astype(np.float64)
         np.savez_compressed(args.dump, **payload)
     return 0
